elections scaffolding, doc config
