+++ .jsdoc.json
@@ -1,6 +1,5 @@
 {
   "opts": {
-    "destination": "doc",
     "recurse": true,
     "access": "all"
   },
