+++ lib/index.js
@@ -30,41 +30,34 @@ const routes = require('./routes');
  * @param {string} config.session.storeFilePath - Path to persistent session
  * storage file.
  *
- * @param {Object} [config.paths] - Endpoint paths.
- * @param {string[]} [config.paths.public=...] - Static content directories.
- * @param {Object} [config.paths.auth] - Authentication-related paths.
- * @param {string} [config.paths.auth.index = '/auth'] - Base path for
- * @param {string} [config.paths.auth.login = ''] - Auth login path.
- * @param {string} [config.paths.auth.logout = '/logout'] - Auth logout path.
- *
  * @returns {express~Application} Resolves with the configured
  * server, or rejects if an error occurred.
  */
 async function server(config) {
     const app = express();
 
-    // Set up default options
     config = Object.assign({
         protocol: 'http',
         hostname: 'localhost',
         port: 8080
     }, config);
 
-    // Merge default paths with configuration
-    config.paths = Object.assign({
-        // Static content paths
+    config.paths = {
         public: [
             path.resolve(__dirname, '../dist'),
             path.resolve(__dirname, '../public')
         ],
 
-        // Authentication-related paths
         auth: {
             index: '/auth',
             login: '/login',
             logout: '/logout'
+        },
+
+        elections: {
+            index: '/elections'
         }
-    }, config.paths);
+    };
 
     await preboots(app, config);
     await middlewares(app, config);
