+++ package.json
@@ -4,8 +4,8 @@
   "description": "CMUJSA website",
   "main": "lib/index.js",
   "scripts": {
-    "doc": "jsdoc -c .jsdoc.json README.md src lib",
-    "doc-watch": "nodemon --config .nodemon.doc.json",
+    "doc-lib": "jsdoc -d doc/lib -c .jsdoc.json README.md lib",
+    "doc-src": "jsdoc -d doc/src -c .jsdoc.json README.md src",
     "live": "node build.js live",
     "watch": "node build.js watch",
     "build": "node build.js",
@@ -51,11 +51,10 @@
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^0.11.2",
     "html-webpack-plugin": "^2.29.0",
-    "jsdoc": "^3.5.3",
+    "jsdoc": "^3.5.5",
     "less": "^2.7.2",
     "less-loader": "^4.0.5",
     "loader-utils": "^1.1.0",
-    "nodemon": "^1.11.0",
     "normalize.css": "^7.0.0",
     "postcss-loader": "^2.0.6",
     "progress": "^2.0.0",
